Finish feature engineering
Two notable changes:

1. Sentence-level labeling. This results in much different data
distributions, thus the parameters of SVM classfiers should be
re-tuned.

2. Add a total of 26 features, documented [1].

Things that you might expect but still missing:
- The data is not balanced
- the SVM classifier is not tuned

Due to the reasons above, the results end result precision and recall
are not good yet. But we are going to move on: feature engineering
results are looking good visually. The only way to further developing
feature is through first moves to classifier tuning, then come back to
debug the feature engineering with feedback from classifier results.

One last notice: no command line interface is maintained at the finest
level, because I don't use that during debugging. The interface shall
be added when other developers are using them, or we are going to ship
to end users. The detailed workflow can be seen in test.py, in
particular two functions:
- further_transform
- build_samples_v2

[1] https://github.com/forrestbao/autoscholar/wiki/Feature-Engineering